+++ runtime/backend/interpreter/src/main/scala/scalascript/interpreter/EvalRuntime.scala
@@ -453,18 +453,18 @@ private[interpreter] object EvalRuntime:
               case other               => eval(other, env, interp)
             (qualC, arg1C, arg2C) match
               case (Pure(qv), Pure(av1), Pure(av2)) =>
-                DispatchRuntime.dispatch(qv, method, av1 :: av2 :: Nil, env, interp)
+                DispatchRuntime.dispatch2(qv, method, av1, av2, env, interp)
               case (Pure(qv), Pure(av1), _) =>
-                FlatMap(arg2C, av2 => DispatchRuntime.dispatch(qv, method, av1 :: av2 :: Nil, env, interp))
+                FlatMap(arg2C, av2 => DispatchRuntime.dispatch2(qv, method, av1, av2, env, interp))
               case (Pure(qv), _, Pure(av2)) =>
-                FlatMap(arg1C, av1 => DispatchRuntime.dispatch(qv, method, av1 :: av2 :: Nil, env, interp))
+                FlatMap(arg1C, av1 => DispatchRuntime.dispatch2(qv, method, av1, av2, env, interp))
               case (Pure(qv), _, _) =>
-                FlatMap(arg1C, av1 => FlatMap(arg2C, av2 => DispatchRuntime.dispatch(qv, method, av1 :: av2 :: Nil, env, interp)))
+                FlatMap(arg1C, av1 => FlatMap(arg2C, av2 => DispatchRuntime.dispatch2(qv, method, av1, av2, env, interp)))
               case (_, Pure(av1), Pure(av2)) =>
-                FlatMap(qualC, qv => DispatchRuntime.dispatch(qv, method, av1 :: av2 :: Nil, env, interp))
+                FlatMap(qualC, qv => DispatchRuntime.dispatch2(qv, method, av1, av2, env, interp))
               case _ =>
                 FlatMap(qualC, qv =>
-                  FlatMap(arg1C, av1 => FlatMap(arg2C, av2 => DispatchRuntime.dispatch(qv, method, av1 :: av2 :: Nil, env, interp))))
+                  FlatMap(arg1C, av1 => FlatMap(arg2C, av2 => DispatchRuntime.dispatch2(qv, method, av1, av2, env, interp))))
           else
             val argComps = argTerms.map {
               case Term.Assign(_, rhs) => eval(rhs, env, interp)